Check and initialise mining fund in chainstate.
While loading the chainstate, check that the mining fund information is
present.  If not, then force the user to -reindex-chainstate.

And if we have a completely fresh chainstate (e. g., when syncing from
scratch or actually reindexing), then we'll initialise the mining fund
with an initial value specified in chain params.  For mainnet, this is
580800, corresponding to the locked premine amount.