Make "zero blockreward height" a chain param.
IXC sets the block reward to zero starting at height 227499.  Instead of
hardcoding this magic number in GetBlockSubsidy, make it a chain param.

This separates the configuration for code clarity, but it also allows us
to set the parameter to 200 for regtest.  With that, we can better test
the mining-fund change later on.