Increase block size to 64 MB at mining-fund fork.
At the fork, also increase the maximum allowed block size to 64 MB
without touching any other parameters.

This is implemented by just increasing MAX_BLOCK_BASE_SIZE, and then
introducing a new check in ContextualCheckBlock, which verifies the old
(stricter) limit prior to the fork height.  That is the cleanest way to
implement things, as it makes it easy to later (after the chain is
beyond the fork) remove the temporary check and end up without any
clutter in the code.